Refactored general searh input into general search container
