--- app/components/GeneralSearchInput/messages.js
@@ -1,13 +1,0 @@
-/*
- * GeneralSearch Messages
- *
- * This contains all the text for the GeneralSearch component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.GeneralSearch.header',
-    defaultMessage: 'This is the GeneralSearch component !',
-  },
-});
